updated city to location, removed additional province from airports, updated the json files,
