+++ src/main/java/com/keyin/rest/airport/Airports.json
@@ -3,231 +3,210 @@
     "airport_ID": 1,
     "code": "YYZ",
     "name": "Toronto Pearson International Airport",
-    "city": {
-      "city_ID": 1,
-      "name": "Toronto",
+    "location": {
+      "location_ID": 1,
+      "city": "Toronto",
       "province": "Ontario"
-    },
-    "province": "Ontario"
+    }
   },
   {
     "airport_ID": 2,
     "code": "YVR",
     "name": "Vancouver International Airport",
-    "city": {
-      "city_ID": 2,
-      "name": "Vancouver",
+    "location": {
+      "location_ID": 2,
+      "city": "Vancouver",
       "province": "British Columbia"
-    },
-    "province": "British Columbia"
+    }
   },
   {
     "airport_ID": 3,
     "code": "YUL",
     "name": "Montréal-Pierre Elliott Trudeau International Airport",
-    "city": {
-      "city_ID": 3,
-      "name": "Montréal",
+    "location": {
+      "location_ID": 3,
+      "city": "Montréal",
       "province": "Quebec"
-    },
-    "province": "Quebec"
+    }
   },
   {
     "airport_ID": 4,
     "code": "YYC",
     "name": "Calgary International Airport",
-    "city": {
-      "city_ID": 4,
-      "name": "Calgary",
+    "location": {
+      "location_ID": 4,
+      "city": "Calgary",
       "province": "Alberta"
-    },
-    "province": "Alberta"
+    }
   },
   {
     "airport_ID": 5,
     "code": "YEG",
     "name": "Edmonton International Airport",
-    "city": {
-      "city_ID": 5,
-      "name": "Edmonton",
+    "location": {
+      "location_ID": 5,
+      "city": "Edmonton",
       "province": "Alberta"
-    },
-    "province": "Alberta"
+    }
   },
   {
     "airport_ID": 6,
     "code": "YOW",
     "name": "Ottawa Macdonald-Cartier International Airport",
-    "city": {
-      "city_ID": 6,
-      "name": "Ottawa",
+    "location": {
+      "location_ID": 6,
+      "city": "Ottawa",
       "province": "Ontario"
-    },
-    "province": "Ontario"
+    }
   },
   {
     "airport_ID": 7,
     "code": "YHZ",
     "name": "Halifax Stanfield International Airport",
-    "city": {
-      "city_ID": 7,
-      "name": "Halifax",
+    "location": {
+      "location_ID": 7,
+      "city": "Halifax",
       "province": "Nova Scotia"
-    },
-    "province": "Nova Scotia"
+    }
   },
   {
     "airport_ID": 8,
     "code": "YQB",
-    "name": "Québec City Jean Lesage International Airport",
-    "city": {
-      "city_ID": 8,
-      "name": "Québec City",
+    "name": "Québec Location Jean Lesage International Airport",
+    "location": {
+      "location_ID": 8,
+      "city": "Québec Location",
       "province": "Quebec"
-    },
-    "province": "Quebec"
+    }
   },
   {
     "airport_ID": 9,
     "code": "YWG",
     "name": "Winnipeg James Armstrong Richardson International Airport",
-    "city": {
-      "city_ID": 9,
-      "name": "Winnipeg",
+    "location": {
+      "location_ID": 9,
+      "city": "Winnipeg",
       "province": "Manitoba"
-    },
-    "province": "Manitoba"
+    }
   },
   {
     "airport_ID": 10,
     "code": "YYJ",
     "name": "Victoria International Airport",
-    "city": {
-      "city_ID": 10,
-      "name": "Victoria",
+    "location": {
+      "location_ID": 10,
+      "city": "Victoria",
       "province": "British Columbia"
-    },
-    "province": "British Columbia"
+    }
   },
   {
     "airport_ID": 11,
     "code": "YLW",
     "name": "Kelowna International Airport",
-    "city": {
-      "city_ID": 11,
-      "name": "Kelowna",
+    "location": {
+      "location_ID": 11,
+      "city": "Kelowna",
       "province": "British Columbia"
-    },
-    "province": "British Columbia"
+    }
   },
   {
     "airport_ID": 12,
     "code": "YXE",
     "name": "Saskatoon John G. Diefenbaker International Airport",
-    "city": {
-      "city_ID": 12,
-      "name": "Saskatoon",
+    "location": {
+      "location_ID": 12,
+      "city": "Saskatoon",
       "province": "Saskatchewan"
-    },
-    "province": "Saskatchewan"
+    }
   },
   {
     "airport_ID": 13,
     "code": "YQR",
     "name": "Regina International Airport",
-    "city": {
-      "city_ID": 13,
-      "name": "Regina",
+    "location": {
+      "location_ID": 13,
+      "city": "Regina",
       "province": "Saskatchewan"
-    },
-    "province": "Saskatchewan"
+    }
   },
   {
     "airport_ID": 14,
     "code": "YXU",
     "name": "London International Airport",
-    "city": {
-      "city_ID": 14,
-      "name": "London",
+    "location": {
+      "location_ID": 14,
+      "city": "London",
       "province": "Ontario"
-    },
-    "province": "Ontario"
+    }
   },
   {
     "airport_ID": 15,
     "code": "YYT",
     "name": "St. John's International Airport",
-    "city": {
-      "city_ID": 15,
-      "name": "St. John's",
+    "location": {
+      "location_ID": 15,
+      "city": "St. John's",
       "province": "Newfoundland and Labrador"
-    },
-    "province": "Newfoundland and Labrador"
+    }
   },
   {
     "airport_ID": 16,
     "code": "YYG",
     "name": "Charlottetown Airport",
-    "city": {
-      "city_ID": 16,
-      "name": "Charlottetown",
+    "location": {
+      "location_ID": 16,
+      "city": "Charlottetown",
       "province": "Prince Edward Island"
-    },
-    "province": "Prince Edward Island"
+    }
   },
   {
     "airport_ID": 17,
     "code": "YFC",
     "name": "Fredericton International Airport",
-    "city": {
-      "city_ID": 17,
-      "name": "Fredericton",
+    "location": {
+      "location_ID": 17,
+      "city": "Fredericton",
       "province": "New Brunswick"
-    },
-    "province": "New Brunswick"
+    }
   },
   {
     "airport_ID": 18,
     "code": "YSJ",
     "name": "Saint John Airport",
-    "city": {
-      "city_ID": 18,
-      "name": "Saint John",
+    "location": {
+      "location_ID": 18,
+      "city": "Saint John",
       "province": "New Brunswick"
-    },
-    "province": "New Brunswick"
+    }
   },
   {
     "airport_ID": 19,
     "code": "YQX",
     "name": "Gander International Airport",
-    "city": {
-      "city_ID": 19,
-      "name": "Gander",
+    "location": {
+      "location_ID": 19,
+      "city": "Gander",
       "province": "Newfoundland and Labrador"
-    },
-    "province": "Newfoundland and Labrador"
+    }
   },
   {
     "airport_ID": 20,
     "code": "YMM",
     "name": "Fort McMurray International Airport",
-    "city": {
-      "city_ID": 20,
-      "name": "Fort McMurray",
+    "location": {
+      "location_ID": 20,
+      "city": "Fort McMurray",
       "province": "Alberta"
-    },
-    "province": "Alberta"
+    }
   },
   {
     "airport_ID": 21,
     "code": "YQT",
     "name": "Thunder Bay International Airport",
-    "city": {
-      "city_ID": 21,
-      "name": "Thunder Bay",
+    "location": {
+      "location_ID": 21,
+      "city": "Thunder Bay",
       "province": "Ontario"
-    },
-    "province": "Ontario"
+    }
   }
 ]
