--- src/main/java/com/keyin/rest/city/CityService.java
@@ -1,4 +1,0 @@
-package com.keyin.rest.city;
-
-public class CityService {
-}
