+++ src/main/java/com/keyin/rest/location/CityService.java
@@ -1,0 +1,4 @@
+package com.keyin.rest.location;
+
+public class CityService {
+}
